requirements.txt
requests
scrapy
asgiref
Django
django-appconf
django-ckeditor
django-debug-toolbar
django-imagekit
django-js-asset
django-mptt
django-recaptcha
django-watson
django-widget-tweaks
Faker
pilkit
Pillow
python-dateutil
pytz
six
sqlparse
text-unidecode
Unidecode
scrapy-proxycrawl-middleware